Initial auth and permissions support

+++ src/controllers/player.js
@@ -1,3 +1,4 @@
+import bcrypt from 'bcrypt'
 import Player from '../models/player'
 
 export const getAllPlayers = async (req, res, next) => {
@@ -29,17 +30,20 @@ export const getPlayerById = async (req, res, next) => {
 export const addPlayer = async (req, res, next) => {
   try {
     const { payload } = req.body
-    const existing = await Player.findOne({ email: payload.email })
+    const { email, password } = payload
+    const existing = await Player.findOne({ email: email })
 
     if (existing) {
       const message = 'Player already exists!'
-      res.json(message)
-      return null
+      return res.status(400).send(message)
     }
 
-    const newPlayer = await Player.create({ ...payload })
+    const salt = await bcrypt.genSalt(10)
+    const passwordHash = await bcrypt.hash(password, salt)
 
-    res.json(newPlayer)
+    const newPlayer = await Player.create({ ...payload, password: passwordHash })
+
+    res.status(200).json(newPlayer)
   } catch (err) {
     console.error(err)
   }
@@ -49,8 +53,17 @@ export const updatePlayer = async (req, res, next) => {
   try {
     const { id } = req.params
     const { payload } = req.body
+    let { password } = payload
+    if (password) {
+      const salt = await bcrypt.genSalt(10)
+      const passwordHash = await bcrypt.hash(password, salt)
+      password = passwordHash
+    }
+    const { isAdmin } = req.user
 
-    let updatedPlayer = await Player.findOneAndUpdate({ _id: id }, { ...payload })
+    if (!isAdmin) return res.status(401).send('Invalid permisssions!')
+
+    let updatedPlayer = await Player.findOneAndUpdate({ _id: id }, { ...payload, password })
 
     updatedPlayer = await Player.findOne({ _id: id })
 
@@ -63,6 +76,10 @@ export const updatePlayer = async (req, res, next) => {
 export const deletePlayer = async (req, res, next) => {
   try {
     const { id } = req.params
+    const { isAdmin } = req.user
+
+    if (!isAdmin) return res.status(401).send('Invalid permisssions!')
+
     const isDeleted = await Player.deleteOne({ _id: id })
 
     if (!isDeleted) throw new Error('Could not delete this id!')
